01 - Reading List: Refactored search feature.
