+++ pruebas/01-reading-list/iivvaannxx/src/features/search/index.ts
@@ -1,13 +1,10 @@
 export { default as SearchBar } from './components/bar.svelte'
+export { performSearch, setSearchQuery } from './state/actions'
+
 export {
 
-  isSearching,
-  setIsSearching,
-
   searchQuery,
-  setSearchQuery,
+  isSearching,
+  shouldSearch
 
-  searchResults,
-  showSearchResults
-
-} from './store'
+} from './state/store'
