+++ pruebas/01-reading-list/iivvaannxx/src/features/search/library/constants.ts
@@ -1,13 +1,11 @@
 import Fuse from 'fuse.js'
-
 import { library } from '$features/library'
-import { setIsSearching } from './store'
 
 /** @brief The minimum number of characters required to start searching. */
 export const MIN_SEARCH_CHARACTERS = 2
 
 /** @brief The options used to configure the fuse instance. */
-const fuseOptions = {
+export const FUSE_OPTIONS = {
 
   ...Fuse.config,
 
@@ -21,17 +19,4 @@ const fuseOptions = {
 }
 
 /** @brief The fuse instance used to search the library. */
-const fuse = new Fuse(library.get(), fuseOptions)
-
-/** @brief Searches the library for books that match the given query. */
-export function search (query: string) {
-
-  // Turn on the searching flag.
-  setIsSearching(true)
-
-  // Start searching and disable it after we are done.
-  const results = fuse.search(query)
-  setIsSearching(false)
-
-  return results
-}
+export const FUSE_INSTANCE = new Fuse(library.get(), FUSE_OPTIONS)
