+++ pruebas/01-reading-list/iivvaannxx/src/features/search/library/utils.ts
@@ -1,0 +1,8 @@
+import { FUSE_INSTANCE } from './constants'
+
+/** @brief Searches the library for the given query. */
+export function search (query: string) {
+
+  const results = FUSE_INSTANCE.search(query)
+  return results.map(result => result.item)
+}
